fix New.pop()
The standard `pop()` implementation inserts the `Pop` instruction right after
the current item, which is wrong in case of `New`. Adding `New` first adds
the `New` item, but then also `Dup`, then constructor arguments, and finally
`Invoke` of the constructor. We need to insert `Pop` after the `Invoke`, to
pop the original instance from the stack _after_ the consturctor finishes.